fraudecero-backend/requirements.txt
Flask
openai>=1.0.0
gunicorn
python-dotenv
flask-cors